Add new Gesture Detector

+++ app/src/main/java/de/sopa/model/StarCalculator.java
@@ -6,7 +6,7 @@ package de.sopa.model;
 public class StarCalculator {
     public int getStars(int neededMoves, int minimumMoves ) {
         int stars = 0;
-        if (minimumMoves / neededMoves > 0.9) {
+        if ((float) (minimumMoves) / neededMoves > 0.9f) {
             stars = 3;
         } else if ((float) (minimumMoves) / neededMoves > 0.4) {
             stars = 2;
